myproject/requirements.txt
biopython
django
requests
psycopg2
pillow